+++ gwcelery/condor.py
@@ -3,6 +3,7 @@ import os
 import shlex
 import subprocess
 import sys
+import time
 
 from celery.bin.base import Command
 import lxml.etree
@@ -17,18 +18,39 @@ def run_exec(*args):
     os.execvp(args[0], args)
 
 
-def submit():
-    """Submit all GWCelery jobs to HTCondor (if not already running)."""
+def running():
+    """Determine if GWCelery is already running under HTCondor."""
     status = subprocess.check_output(('condor_q', '-xml') + CONSTRAINTS)
     classads = lxml.etree.fromstring(status)
-    if classads.find('.//c') is None:
-        run_exec('condor_submit', SUBMIT_FILE)
-    else:
+    return classads.find('.//c') is not None
+
+
+def submit():
+    """Submit all GWCelery jobs to HTCondor (if not already running)."""
+    if running():
         print('error: GWCelery jobs are already running.\n'
               'You must first remove exist jobs with "gwcelery condor rm".\n'
               'To see the status of those jobs, run "gwcelery condor q".',
               file=sys.stderr)
         sys.exit(1)
+    else:
+        run_exec('condor_submit', SUBMIT_FILE)
+
+
+def resubmit():
+    """Remove any running GWCelery jobs and resubmit to HTCondor."""
+    if running():
+        subprocess.check_call(('condor_rm',) + CONSTRAINTS)
+    timeout = 60
+    start = time.monotonic()
+    while time.monotonic() - start < timeout:
+        if not running():
+            break
+        time.sleep(1)
+    else:
+        print('error: Could not stop all GWCelery jobs', file=sys.stderr)
+        sys.exit(1)
+    run_exec('condor_submit', SUBMIT_FILE)
 
 
 def rm():
@@ -57,7 +79,7 @@ class CondorCommand(Command):
 
     def add_arguments(self, parser):
         subparsers = parser.add_subparsers()
-        for func in [submit, rm, hold, release, q]:
+        for func in [submit, rm, hold, release, resubmit, q]:
             subparser = subparsers.add_parser(func.__name__, help=func.__doc__)
             subparser.set_defaults(func=func)