+++ CHANGES.md
@@ -7,6 +7,10 @@
 -   Add `gwcelery condor resubmit` as a shortcut for
     `gwcelery condor rm; gwcelery condor submit`.
 
+-   Fix deprecation warning due to renaming of
+    `ligo.gracedb.rest.Gracedb.createTag` to
+    `ligo.gracedb.rest.Gracedb.addTag`.
+
 ## 0.0.28 (2018-07-25)
 
 -   Add injection checks to `check_vector`.